+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- [spatie/laravel-activitylog](https://github.com/spatie/laravel-activitylog)
+   - ``.env``: ``ACTIVITY_LOGGER_ENABLED``, default ``true``
+
 ## [v1.0.0](https://gitlab.com/daveroverts/Book-me-a-cookie/compare/v0.7.0c...v1.0.0) - 2018-12-15
 
 ### Added